% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_moving_average.R
\name{add_moving_average}
\alias{add_moving_average}
\title{Add Moving Average for an Arbitrary Number of Intervals}
\usage{
add_moving_average(
  .data,
  sort_cols,
  val,
  intervals = 2,
  res_val = "{{val}}_mavg",
  restore_order = FALSE
)
}
\arguments{
\item{.data}{A tibble or data frame.}

\item{sort_cols}{Columns used for sorting passed in a manner consistent with
\code{\link[dplyr]{arrange}}}

\item{val}{Column used to calculate moving average passed as bare column
name or a character string.}

\item{res_val}{Resulting moving average, defaults to name of \code{val}
suffixed with \code{_mavg}.}

\item{restore_order}{A logical, defaults to \code{FALSE} if \code{TRUE} it
will restore original data order.}
}
\value{
A tibble with appended moving average.
}
\description{
The functions adds moving average for an arbitrary number of intervals. The
  data can be returned sorted or according to the original order.
}
\details{
The function can be used independently or within dplyr pipeline.
}
\examples{
add_moving_average(mtcars, sort_cols = c("mpg", "cyl"), val = hp, intervals = 2)
}
