% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-sql-file.R
\name{clean_sql_file}
\alias{clean_sql_file}
\title{Clean SQL File}
\usage{
clean_sql_file(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Cleans SQL file removing comments and collapsing all the lines together.
}
