% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-missing.R
\name{compare_missing}
\alias{compare_missing}
\title{Compare Missing Values Across Vectors}
\usage{
compare_missing(..., add_percentage = FALSE)
}
\arguments{
\item{...}{Vectors to compare.}

\item{add_percentage}{A logical, defaults to \code{FALSE}. If \code{TRUE} will
add a column with percentage for missing values.}
}
\value{
A tibble
}
\description{
The function generates a simple data frame outlining a number of missing
  values across provided vectors
}
\examples{
vecA <- 1:10
vecB <- c(1:100, NA_integer_)
vecC <- LETTERS
compare_missing(vecA, vecB, vecC)
}
