% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_multiple_vectors.R
\name{compare_multiple_vectors}
\alias{compare_multiple_vectors}
\title{Compare Values of Multiple Vectors}
\usage{
compare_multiple_vectors(x, ...)
}
\arguments{
\item{x}{Alist of vectors to compare}

\item{...}{as in  \code{\link[base]{all.eqal}}}
}
\value{
A logical
}
\description{
The function compares values across multiple vectors using
  \code{\link[base]{all.eqal}}.
}
\examples{
# Returns TRUE
compare_multiple_vectors(c(1,1,1), c(1,1,1))
# Returns FALSE
compare_multiple_vectors(c(1,1,1), c(1,1,1), c(1,2,1))
# Returns FALSE
compare_multiple_vectors(c(1,2,3), c(3,2,1))
}
