% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_three.R
\name{compare_three}
\alias{compare_three}
\title{Compare Three Vectors}
\usage{
compare_three(
  x,
  y,
  z,
  compare_fun = "identical",
  result = c("single", "itemized")
)
}
\arguments{
\item{x, y, z}{Vectors of equal length to compare}

\item{compare_fun}{A function used to compare defaults to
\code{\link[base]{identical}}.}

\item{result}{For \emph{single} it will return logical scalar \code{TRUE} if
all arguments are identical; for \emph{itemized} it will return vector of
length equal to lengths of passed arguments. Defaults to \emph{single}.}
}
\value{
A logical vector of length equal to \code{a, b, c}
}
\description{
Convenience function comparing three elements using a desired
  comparator function.
}
\details{
The results of the comparator function are tested with
}
\examples{
compare_three(1:10, 1:10, 1:10)
compare_three(c("b", "a"), c("b", "b"), c("b", "c"), result = "itemized")
}
