% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_triangle.R
\name{get_triangle}
\alias{get_triangle}
\title{Get Lower or Upper Triangle of Correlations Matrix}
\usage{
get_triangle(cormat, part = "upper")
}
\arguments{
\item{cormat}{Usually correlation matrix}

\item{part}{Single string vector decide whether to return \code{upper} or
\code{lower} part.}
}
\value{
Object of dimensions \code{dim(cormat)} with \code{NA} values being
  present in lower or upper triangle.
}
\details{
The functions as a wrapper on \code{\link[base]{lower.tri}} and
  \code{\link[base]{upper.tri}} functions.
}
\examples{
M <- matrix(1:20, 5, 5)
get_triangle(cormat = M, part = "upper")
}
