% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeway.R
\name{\%<=>\%}
\alias{\%<=>\%}
\title{Three-way operator}
\usage{
lhs \%<=>\% rhs
}
\arguments{
\item{lhs}{Left hand side operator}

\item{rhs}{Rand hand side operator}
}
\value{
Greater value or zero if equal.
}
\description{
A three-way comparison takes two values A and B determines
  whether A < B, A = B, or A > B in a single operation.
}
\examples{
1 \%<=>\% 2
2 \%<=>\% 1
3 \%<=>\% 3
}
