% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{guess_vector_type}
\alias{guess_vector_type}
\title{Guess Vector Type}
\usage{
guess_vector_type(x)
}
\arguments{
\item{x}{A vector of factor class.}
}
\value{
A scalar string with class name.
}
\description{
Function analyses content of a factor vector and attempts to
guess the correct type.
}
\section{Acknowledgements}{

The initial version of this function was posted and discussed on
\href{https://stackoverflow.com/q/52943081/1655567}{StackOverflow}.
}

