% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEmisc.R
\docType{package}
\name{KEmisc-package}
\alias{KEmisc}
\alias{KEmisc-package}
\title{Kmisc: colleaction of misceallous functions}
\description{
Kmisc provides a disorganised collection of convenience and
  utility functions.
}
\details{
The package was developed out of need to keep random handy functions,
 which don't appear to belong anywhere in one place.
}
\section{Acknowledgements}{

The following package is developed in the open with frequent use of
  StackOverflow. I've strived to reference relevant discussions in the
  code and acknowledgements sections. Ultimately, all relevant discussion
  can be found back via my
  \href{https://stackoverflow.com/users/1655567/konrad?tab=questions}{SO profile.}
}

\author{
\strong{Maintainer}: Konrad Zdeb \email{konrad.zdeb@me.com}

}
