% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_padded_tibble.R
\name{make_padded_tibble}
\alias{make_padded_tibble}
\title{Make Padded Tibble}
\usage{
make_padded_tibble(lst)
}
\arguments{
\item{lst}{A list of vector to combine}
}
\value{
A tibble
}
\description{
Efficiently combine vectors of unequal lengths into a tibble
}
\details{
The function is inspired by this answer given on StackOverflow
  \url{https://stackoverflow.com/a/7196683/1655567}
}
\examples{
make_padded_tibble(list(vecA = 1:10, vecB = 1:5))
}
