% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_recent.R
\name{most_recent}
\alias{most_recent}
\title{Find most recent file or folder}
\usage{
most_recent(search_path, type = c("any", "file", "directory", "symlink"))
}
\arguments{
\item{search_path}{A path to search}

\item{type}{Type of an object to return. Defaults to \code{any}, available
options correspond to \code{type} argument in
\code{\link[fs:dir_ls]{fs::dir_ls}}.}
}
\value{
A \code{\link[fs]{path}} object.
}
\description{
Searches path provided in the \code{search_path} and returns most recent file
  or directory.
}
\details{
The function makes use of the directory listing function available
  in the \code{fs} package - \code{\link[fs:dir_ls]{fs::dir_ls}} and
  consequently can return any objects that
  \code{\link[fs:dir_ls]{fs::dir_ls}} can identify.
}
\examples{
most_recent(tempdir())
}
