% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving-average.R
\name{moving_average}
\alias{moving_average}
\title{Add Moving Average for an Arbitrary Number of Intervals}
\usage{
moving_average(x, x_order, interval, na.rm = FALSE)
}
\arguments{
\item{x}{A vector used to calculate moving average.}

\item{x_order}{A vector used for sorting of \code{x} defaults to
\code{sort(x)}.}

\item{interval}{A number of intervals for moving average, defaults to
\code{round(length(x) / 3, 0)}.}

\item{na.rm}{As in \code{\link[base]{sum}}, defaults to \code{FALSE}.}
}
\value{
A vector of length equivalent to \code{x} with \code{NA} for
  n < interval as in \code{x[n]}.
}
\description{
The functions adds moving average for an arbitrary number of intervals. The
  data can be returned sorted or according to the original order.
}
\details{
The function can be used independently or within dplyr pipeline.
}
\examples{
moving_average(x = sample(1:100, 10), x_order = 10:1)
}
