% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_cuts.R
\name{pretty_cuts}
\alias{pretty_cuts}
\title{Pretty Cuts}
\usage{
pretty_cuts(cut_str)
}
\arguments{
\item{cut_str}{A character or factorvector produced by \code{\link[base]{cut}}
or by \code{\link[Hmisc]{cut2}} or a similar function.}
}
\value{
A character vector.
}
\description{
Function removing set notation used by \code{\link[base]{cut}}
  or by \code{\link[Hmisc]{cut2}}.
}
\details{
For a vector with sets \code{(2e+03,4e+03]}
  \code{pretty_cuts} will return  \code{"2.01 - 3"}.
}
\examples{
set.seed(123)
pretty_cuts(cut(x = runif(n = 1e3, min = 1, 1e4), breaks = 5))

}
