% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_table.R
\name{preview_table}
\alias{preview_table}
\title{Preview Table}
\usage{
preview_table(x, Show = TRUE)
}
\arguments{
\item{x}{Vector passed to \code{\link[base]{table}}.}

\item{Show}{A logical, defaults to \code{\link[base]{TRUE}}, runs
\code{\link[utils]{View}}.}
}
\value{
A data frame.
}
\description{
The function constructs a data frame from \code{\link[base]{table}}
  function and optionally passes results to \code{\link[utils]{View}}
  function for easier preview.
}
\examples{
preview_table(x = c(rep(1, 10), rep(5, 2)))
}
