% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-all-objects.R
\name{save_all_objects}
\alias{save_all_objects}
\title{Save All Objects to RDS Files}
\usage{
save_all_objects(
  env = parent.frame(n = 1),
  output_path = tempdir(),
  append_timestamps = TRUE,
  skip_functions = TRUE,
  ...
)
}
\arguments{
\item{env}{Environment to search for objects. Defaults to
\code{\link[base:parent.frame]{parent_frame(n = 1)}}.}

\item{output_path}{Defaults to \code{\link[base]{tempdir}}.}

\item{append_timestamps}{A boolean, defaults to \code{TRUE} will append
timestamps to each object name.}

\item{skip_functions}{Defaults to \code{TRUE}, removes functions from objects
returned by \code{\link[base]{ls}}.}

\item{...}{As in \code{\link[base]{saveRDS}}.}
}
\value{
A character vector of absolute file paths
}
\description{
Iterate over objects available in the designated environment and create a RDS
  object for each separate object.
}
\details{
This a convenience function iterates over objects available within
  the designated environment and creates a RDS file for each single item. The
  files can have automatically dates added to them.
}
\examples{
a <- 1
save_all_objects()
}
