% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_without_zeros.R
\name{units_without_zero}
\alias{units_without_zero}
\title{Produce Formatted Unit Without Zero}
\usage{
units_without_zero(x, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{...}{other arguments passed to \code{\link[scales]{unit_format}}}
}
\value{
A character vector.
}
\description{
Conveniance wrapper on useful \code{\link[scales]{unit_format}}
  function offered within \code{scales} package. Useful when producing scales
  without 0: \code{"" "1K" "10K"} instead of \code{"0k" "1K" "10K"}.
}
\examples{
vec_tst <- c(0, 1000, 1e6)
units_without_zero(vec_tst, unit = "K", scale = 1e-3, sep = "")
# "0"      "1K"     "1,000K"
# For comparison
scales::unit_format(unit = "K", scale = 1e-3, sep = "")(vec_tst)
# [1] "0K"     "1K"     "1,000K"

}
