% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dated_name.R
\name{create_dated_file_name}
\alias{create_dated_file_name}
\title{Create Dated File/Path Name}
\usage{
create_dated_file_name(x, timestamp_format = "\%d-\%m-\%Y", sep = "-")
}
\arguments{
\item{x}{A character vector.}

\item{timestamp_format}{A string scalar.}

\item{sep}{A string scalar.}
}
\value{
A character vector of length corresponding to x.
}
\description{
This is trival function that wraps \code{\link[base]{paste}}
  call and prduces a string with \code{\link[base]{Sys.time}} results.
}
\examples{
create_dated_file_name(x = "file_name")
}
