% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmisc.R
\docType{package}
\name{Kmisc}
\alias{Kmisc}
\alias{Kmisc-package}
\title{Kmisc: colleaction of misceallous functions}
\description{
Kmisc provides a disorganised collection of conveniance and utility functions.
}
