% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_RDS_to_global.R
\name{read_RDS_to_global}
\alias{read_RDS_to_global}
\title{Read RDS Ffile Directly to Global Environment}
\usage{
read_RDS_to_global(file_path, verbose = TRUE)
}
\arguments{
\item{verbose}{A logica, if \code{TRUE}}
}
\value{
A R object corresponding to \code{basename} on \code{file_path}
  argument.
}
\description{

}
\examples{
\dontrun{
  tmpCars <- tempfile(pattern = "mtcars_", fileext = ".RDS")
  saveRDS(object = mtcars, file = tmpCars)
  read_RDS_to_global(tmpCars)
}

}
